aws/Lambda/Lambda-monitor/scripts/cloudwatch_alarm/requirements.txt
pyyaml==6.0.1
boto3==1.28.34